# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Resource-Service is service for storing kubernetes resources in Mongo DB.",
    "title": "Resource-Service",
    "version": "0.0.1"
  },
  "paths": {
...
    "/namespaces/{namespace}/deployments": {
      "get": {
        "tags": [
          "Deployment"
        ],
        "summary": "Get deployments list.",
        "operationId": "GetDeploymentsListHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "deployments list",
            "schema": {
              "$ref": "#/definitions/DeploymentsResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Deployment"
        ],
        "summary": "Create deployment.",
        "operationId": "CreateDeploymentHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "deployment created",
            "schema": {
              "$ref": "#/definitions/ResourceDeploy"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "Deployment"
        ],
        "summary": "Delete all deployments in namespace.",
        "operationId": "DeleteAllDeploymentsHandler",
        "parameters": [
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "all deployments in namespace deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "private"
      }
    },
    "/namespaces/{namespace}/deployments/{deployment}": {
      "get": {
        "tags": [
          "Deployment"
        ],
        "summary": "Get deployment active version.",
        "operationId": "GetActiveDeploymentHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "deployment",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "deployment",
            "schema": {
              "$ref": "#/definitions/ResourceDeploy"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "put": {
        "tags": [
          "Deployment"
        ],
        "summary": "Update deployment.",
        "operationId": "UpdateDeployment",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "deployment",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "deployment updated",
            "schema": {
              "$ref": "#/definitions/ResourceDeploy"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "Deployment"
        ],
        "summary": "Delete deployment.",
        "operationId": "DeleteDeploymentHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "deployment",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "deployment deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/namespaces/{namespace}/deployments/{deployment}/image": {
      "put": {
        "tags": [
          "Deployment"
        ],
        "summary": "Update image in deployments container.",
        "operationId": "SetContainerImageHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "deployment",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateImage"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "deployment updated",
            "schema": {
              "$ref": "#/definitions/ResourceDeploy"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/namespaces/{namespace}/deployments/{deployment}/replicas": {
      "put": {
        "tags": [
          "Deployment"
        ],
        "summary": "Update deployments replicas count.",
        "operationId": "SetReplicasHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "deployment",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateReplicas"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "deployment updated",
            "schema": {
              "$ref": "#/definitions/ResourceDeploy"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/namespaces/{namespace}/deployments/{deployment}/versions": {
      "get": {
        "tags": [
          "Deployment"
        ],
        "summary": "Get deployments list.",
        "operationId": "GetDeploymentVersionsListHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "deployment",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "deployment versions list",
            "schema": {
              "$ref": "#/definitions/DeploymentsResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/namespaces/{namespace}/deployments/{deployment}/versions/{version}/diff": {
      "post": {
        "tags": [
          "Deployment"
        ],
        "summary": "Compare deployment versions with previous version.",
        "operationId": "DiffDeploymentPreviousVersionsHandler",
        "parameters": [
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "deployment",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "version",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "diff"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "private"
      }
    },
    "/namespaces/{namespace}/deployments/{deployment}/versions/{version}/diff/{version2}": {
      "post": {
        "tags": [
          "Deployment"
        ],
        "summary": "Compare two deployment versions.",
        "operationId": "DiffDeploymentVersionsHandler",
        "parameters": [
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "deployment",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "version",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "version2",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "diff"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "private"
      }
    },
...
    "/namespaces/{namespace}/solutions/{solution}/deployments": {
      "delete": {
        "tags": [
          "Service"
        ],
        "summary": "Delete all solution deployments.",
        "operationId": "DeleteAllSolutionDeploymentsHandler",
        "parameters": [
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "all solution deployments deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "private"
      }
    },
...
  },
  "definitions": {
...
    "DeploymentsResponse": {
      "description": "DeploymentsResponse -- deployments response",
      "type": "object",
      "properties": {
        "deployments": {
          "$ref": "#/definitions/ListDeploy"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/deployment"
    },
...
    "ListDeploy": {
      "description": "Deployment -- deployments list",
      "type": "array",
      "items": {
        "$ref": "#/definitions/ResourceDeploy"
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/deployment"
    },
...
    "UpdateImage": {
      "description": "UpdateImage -- model for update container image request",
      "type": "object",
      "required": [
        "container_name",
        "image"
      ],
      "properties": {
        "container_name": {
          "type": "string",
          "x-go-name": "Container"
        },
        "image": {
          "type": "string",
          "x-go-name": "Image"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "UpdateReplicas": {
      "description": "UpdateReplicas -- contains new number of replicas",
      "type": "object",
      "required": [
        "replicas"
      ],
      "properties": {
        "replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Replicas"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "errOAIGen": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "description": "Fields -- represents additional key-value fields of error",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "id": {
          "$ref": "#/definitions/errOaiGenId"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "github.com/containerum/cherry"
    },
...
  },
...
}